git-svn-id: https://dev.wibe.media/svn/code/java/trunk/litiengine@1983 16493bb7-3577-b145-b9c3-72a97d279fb6

+++ src/de/gurkenlabs/litiengine/graphics/particles/emitters/ShimmerEmitter.java
@@ -28,7 +28,7 @@ public class ShimmerEmitter extends Emitter {
    * @param originY
    *          the origin y
    */
-  public ShimmerEmitter(final int originX, final int originY) {
+  public ShimmerEmitter(final double originX, final double originY) {
     super(originX, originY);
   }
